Making gt-referencing OSGi compliant.
Updated poms for gt-referencing and its dependency, gt-metadata, to
include the maven-bundle-plugin to provide additional OSGi compliancy to
this fork.

Additionally, added a comment to the description element of the opengis
module to identify it as OSGi compliant.

Finally, removed an image file from the docs that was removed in the
originating project from which this project is forked.